@@ -1,54 +1,38 @@
 source 'https://rubygems.org'
 
+# Railsのバージョンを更新
+gem 'rails', '~> 6.1.4'
+# sqlite3のバージョンを更新
+gem 'sqlite3', '~> 1.4'
+# sass-railsのバージョンを更新
+gem 'sass-rails', '~> 6.0'
+# Uglifierのバージョンを更新
+gem 'uglifier', '>= 4.2.0'
+# CoffeeScriptのバージョンを更新
+gem 'coffee-rails', '~> 5.0'
+# jQueryは変更なし
+gem 'jquery-rails', '~> 4.4'  # 互換性のあるバージョンに更新
+# Turbolinksのバージョンを更新
+gem 'turbolinks', '~> 5.2'
+# jbuilderのバージョンを更新
+gem 'jbuilder', '~> 2.11'
+# sdocのバージョンを更新（ドキュメント生成用）
+gem 'sdoc', '~> 1.1.0', group: :doc
 
-# Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.2.6'
-# Use sqlite3 as the database for Active Record
-gem 'sqlite3'
-# Use SCSS for stylesheets
-gem 'sass-rails', '~> 5.0'
-# Use Uglifier as compressor for JavaScript assets
-gem 'uglifier', '>= 1.3.0'
-# Use CoffeeScript for .coffee assets and views
-gem 'coffee-rails', '~> 4.1.0'
-# See https://github.com/rails/execjs#readme for more supported runtimes
-# gem 'therubyracer', platforms: :ruby
-
-# Use jquery as the JavaScript library
-gem 'jquery-rails'
-# Turbolinks makes following links in your web application faster. Read more: https://github.com/rails/turbolinks
-gem 'turbolinks'
-# Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
-gem 'jbuilder', '~> 2.0'
-# bundle exec rake doc:rails generates the API under doc/api.
-gem 'sdoc', '~> 0.4.0', group: :doc
-
-gem 'acts-as-taggable-on', '~> 3.5.0'
+# その他のgemもバージョンを更新
+gem 'acts-as-taggable-on', '~> 7.0'  # 互換性のある新しいバージョンに更新
 gem 'acts_as_follower', '~> 0.2.1'
-gem 'devise', '~> 4.2.0'
-gem 'jwt', '~> 1.5.4'
-gem 'rack-cors', '~> 0.4.0'
+gem 'devise', '~> 4.8'
+gem 'jwt', '~> 2.2'
+gem 'rack-cors', require: 'rack/cors'
 gem 'puma'
 
-# Use ActiveModel has_secure_password
-# gem 'bcrypt', '~> 3.1.7'
-
-# Use Unicorn as the app server
-# gem 'unicorn'
-
-# Use Capistrano for deployment
-# gem 'capistrano-rails', group: :development
-
 group :development, :test do
-  # Call 'byebug' anywhere in the code to stop execution and get a debugger console
   gem 'byebug'
 end
 
 group :development do
-  # Access an IRB console on exception pages or by using <%= console %> in views
-  gem 'web-console', '~> 2.0'
-
-  # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
+  gem 'web-console', '~> 4.1'
   gem 'spring'
 end
 
