@@ -1,4 +1,4 @@
-class AddProfileFieldsToUsers < ActiveRecord::Migration
+class AddProfileFieldsToUsers < ActiveRecord::Migration[6.1]
   def change
     add_column :users, :username, :string
     add_index :users, :username, unique: true
