@@ -1,4 +1,4 @@
-class CreateArticles < ActiveRecord::Migration
+class CreateArticles < ActiveRecord::Migration[6.1]
   def change
     create_table :articles do |t|
       t.string :title
