@@ -1,5 +1,5 @@
 # This migration comes from acts_as_taggable_on_engine (originally 4)
-class AddMissingTaggableIndex < ActiveRecord::Migration
+class AddMissingTaggableIndex < ActiveRecord::Migration[6.1]
   def self.up
     add_index :taggings, [:taggable_id, :taggable_type, :context]
   end
