@@ -1,4 +1,4 @@
-class CreateFavorites < ActiveRecord::Migration
+class CreateFavorites < ActiveRecord::Migration[6.1]
   def change
     create_table :favorites do |t|
       t.references :user, index: true, foreign_key: true
