@@ -1,4 +1,4 @@
-class ActsAsFollowerMigration < ActiveRecord::Migration
+class ActsAsFollowerMigration < ActiveRecord::Migration[6.1]
   def self.up
     create_table :follows, force: true do |t|
       t.references :followable, polymorphic: true, null: false
